Move home components to route group
