+++ app/(home)/ProjectIllustration.tsx
