+++ app/(home)/ProjectIllustrationClickArea.tsx
