+++ app/(home)/_intro/CareerParagraph.tsx
