+++ app/(home)/_intro/EducationParagraph.tsx
