+++ app/(home)/_intro/RelocationParagraph.tsx
