+++ app/(home)/characterCountContext.tsx
