+++ app/(home)/projectCardsInfo.tsx
@@ -1,38 +1,38 @@
 import JourneyFinderSvg, {
   SVGRComponent,
-} from './projectCardSvgs/1_JourneyFinder_Default.svg';
-import JourneyFinderSvgHighlight from './projectCardSvgs/1_JourneyFinder_Hover.svg';
-import JourneyFinderSvgOutline from './projectCardSvgs/1_JourneyFinder_Outline.svg';
-import JourneyFinderSvgGuide from './projectCardSvgs/1_JourneyFinder_Guide.svg';
-import VijksPt1Svg from './projectCardSvgs/2_Vijks_Pt1_Default.svg';
-import VijksPt1SvgHighlight from './projectCardSvgs/2_Vijks_Pt1_Hover.svg';
-import VijksPt1SvgOutline from './projectCardSvgs/2_Vijks_Pt1_Outline.svg';
-import VijksPt1SvgGuide from './projectCardSvgs/2_Vijks_Pt1_Guide.svg';
-import VijksPt2Svg from './projectCardSvgs/2_Vijks_Pt2_Default.svg';
-import VijksPt2SvgHighlight from './projectCardSvgs/2_Vijks_Pt2_Hover.svg';
-import VijksPt2SvgOutline from './projectCardSvgs/2_Vijks_Pt2_Outline.svg';
-import VijksPt2SvgGuide from './projectCardSvgs/2_Vijks_Pt2_Guide.svg';
-import BrandingSvg from './projectCardSvgs/3_Branding_Default.svg';
-import BrandingSvgHighlight from './projectCardSvgs/3_Branding_Hover.svg';
-import BrandingSvgOutline from './projectCardSvgs/3_Branding_Outline.svg';
-import BrandingSvgGuide from './projectCardSvgs/3_Branding_Guide.svg';
+} from './_projectCardSvgs/1_JourneyFinder_Default.svg';
+import JourneyFinderSvgHighlight from './_projectCardSvgs/1_JourneyFinder_Hover.svg';
+import JourneyFinderSvgOutline from './_projectCardSvgs/1_JourneyFinder_Outline.svg';
+import JourneyFinderSvgGuide from './_projectCardSvgs/1_JourneyFinder_Guide.svg';
+import VijksPt1Svg from './_projectCardSvgs/2_Vijks_Pt1_Default.svg';
+import VijksPt1SvgHighlight from './_projectCardSvgs/2_Vijks_Pt1_Hover.svg';
+import VijksPt1SvgOutline from './_projectCardSvgs/2_Vijks_Pt1_Outline.svg';
+import VijksPt1SvgGuide from './_projectCardSvgs/2_Vijks_Pt1_Guide.svg';
+import VijksPt2Svg from './_projectCardSvgs/2_Vijks_Pt2_Default.svg';
+import VijksPt2SvgHighlight from './_projectCardSvgs/2_Vijks_Pt2_Hover.svg';
+import VijksPt2SvgOutline from './_projectCardSvgs/2_Vijks_Pt2_Outline.svg';
+import VijksPt2SvgGuide from './_projectCardSvgs/2_Vijks_Pt2_Guide.svg';
+import BrandingSvg from './_projectCardSvgs/3_Branding_Default.svg';
+import BrandingSvgHighlight from './_projectCardSvgs/3_Branding_Hover.svg';
+import BrandingSvgOutline from './_projectCardSvgs/3_Branding_Outline.svg';
+import BrandingSvgGuide from './_projectCardSvgs/3_Branding_Guide.svg';
 // Reuse Outline for Nasa SVG
-// import NasaSvg from './projectCardSvgs/4_Nasa_Default.svg';
-import NasaSvgHighlight from './projectCardSvgs/4_Nasa_Hover.svg';
-import NasaSvgOutline from './projectCardSvgs/4_Nasa_Outline.svg';
-import NasaSvgGuide from './projectCardSvgs/4_Nasa_Guide.svg';
-import InteractivesPt1Svg from './projectCardSvgs/5_Interactives_Pt1_Default.svg';
-import InteractivesPt1SvgHighlight from './projectCardSvgs/5_Interactives_Pt1_Hover.svg';
-import InteractivesPt1SvgOutline from './projectCardSvgs/5_Interactives_Pt1_Outline.svg';
-import InteractivesPt1SvgGuide from './projectCardSvgs/5_Interactives_Pt1_Guide.svg';
-import InteractivesPt2Svg from './projectCardSvgs/5_Interactives_Pt2_Default.svg';
-import InteractivesPt2SvgHighlight from './projectCardSvgs/5_Interactives_Pt2_Hover.svg';
-import InteractivesPt2SvgOutline from './projectCardSvgs/5_Interactives_Pt2_Outline.svg';
-import InteractivesPt2SvgGuide from './projectCardSvgs/5_Interactives_Pt2_Guide.svg';
-import VisualStoriesSvg from './projectCardSvgs/6_VisualStories_Default.svg';
-import VisualStoriesSvgHighlight from './projectCardSvgs/6_VisualStories_Hover.svg';
-import VisualStoriesSvgOutline from './projectCardSvgs/6_VisualStories_Outline.svg';
-import VisualStoriesSvgGuide from './projectCardSvgs/6_VisualStories_Guide.svg';
+// import NasaSvg from './_projectCardSvgs/4_Nasa_Default.svg';
+import NasaSvgHighlight from './_projectCardSvgs/4_Nasa_Hover.svg';
+import NasaSvgOutline from './_projectCardSvgs/4_Nasa_Outline.svg';
+import NasaSvgGuide from './_projectCardSvgs/4_Nasa_Guide.svg';
+import InteractivesPt1Svg from './_projectCardSvgs/5_Interactives_Pt1_Default.svg';
+import InteractivesPt1SvgHighlight from './_projectCardSvgs/5_Interactives_Pt1_Hover.svg';
+import InteractivesPt1SvgOutline from './_projectCardSvgs/5_Interactives_Pt1_Outline.svg';
+import InteractivesPt1SvgGuide from './_projectCardSvgs/5_Interactives_Pt1_Guide.svg';
+import InteractivesPt2Svg from './_projectCardSvgs/5_Interactives_Pt2_Default.svg';
+import InteractivesPt2SvgHighlight from './_projectCardSvgs/5_Interactives_Pt2_Hover.svg';
+import InteractivesPt2SvgOutline from './_projectCardSvgs/5_Interactives_Pt2_Outline.svg';
+import InteractivesPt2SvgGuide from './_projectCardSvgs/5_Interactives_Pt2_Guide.svg';
+import VisualStoriesSvg from './_projectCardSvgs/6_VisualStories_Default.svg';
+import VisualStoriesSvgHighlight from './_projectCardSvgs/6_VisualStories_Hover.svg';
+import VisualStoriesSvgOutline from './_projectCardSvgs/6_VisualStories_Outline.svg';
+import VisualStoriesSvgGuide from './_projectCardSvgs/6_VisualStories_Guide.svg';
 import { WrittenProject } from '@/projectContents';
 
 type X = number;
