+++ app/(home)/usePercentCharactersFlipped.ts
